Dropdown, Settings page, Logout
Logout now works. When a person presses sign out on the dropdown it will send them to logout.php and then destroy the session and send them to login.php

Dropdown is finally done. I don't know why I didn't do it sooner but the frontend for the dropdown is complete.

The users information now displays around the website including their full name and email address. This is used on the dashboard, navbar and dropdown.

Settings page frontend is done. I'm not the most happy with the design but it'll do. The backend isn't implemented as I'm struggling with some parts but it'll be done today.