Admin, registration bug and full name.
Dropdown now has "admin" which will be used to access the admin panel.

First Name & Last Name are now collected on the registration page.

I made a mistake on line 26 on the register.php. I forgot a bracket after password which means that it as never checking correctly. It is fixed now.

Also fixed links on other authentication pages.