Emails!
This is the start of implementing emails into the billing system.

So far there isn't much functionality within it.

However, after a user registers it will now send them an email welcoming them to the website. This will be used to make sure that the users confirm their account upon sign up.

The rest is just creating a main table for the main site settings and to implement SMTP.