Refacorting
So, I took a break from working on projects to learn more about PHP and best practises. I think at this point this entire code base needs to be worked.

So I have started to do so. You can see the changes but here are a few so far;

Having a views files and splitting the HTML and PHP completely so they're in completely different sections.

Not storing all objects into a variable when I wont be using the variable further in the code.

Fixing some QOL things such as the fact on session it was re-directing to some random directory that doesn't exist in these files.

Disabling debug mode on mail so it stops printing the entire mail information