File clean up
More updates to the software to extend the revamp even further.

Started by organising the entire assets folder a bit more and having the icons with correct names in correct folders to make it easier to understand

Htaccess has now been added so we can have /register instead of /register.php and so fourth.

Issue in cpanel class has now been fixed, stupid typo.

Edited session to fix a bug and make it look easier to read.

Dashboard meta file was now created to shorten the files of them all. Dashboard navbar also moved so it's not in the individual files. Footer also generated for the tiny bit at the bottom of the page.

Fixed some math on the service page as the calculation was just incorrect.

CSS has also now been correctly minified then the other half arsed parts that weren't and were.