Big overhaul of code.
Correctly send the URL for the web page to the person who is confirming. I stupidly before put this as my personal localhost which wouldn't work for anyone else.

In the last few pushes I think that the confirm view was saved as the forgot password view but this has now been fixed.

We now also check if the ID is set and if its empty before executing anything else on the service page. The code is also much easier to read and manage.

Database class cleaned up a bit more and same with the confirm class.

Updating sanitization of input to php 8.1

Login class has been updated to be more efficient and less lines of code.

User can not login anymore if the user is banned.

Register class has been updated a bit. A lot of variables and empty stupid lines removed to make it shorter.

Removed empty variable $code in the email function that was doing literally nothing.