More updates
Fixing settings so it actually functions and works as well as it should. Still need to add password reset. Also changed firstName -> firstname and lastName -> lastname so it is easier to manage.

Register works I wasn't checking if the email matched so it could make as many accounts with one email, it's fixed now.

Cleaned up confirm so it's safer and easier to manage.

CPanel class has been overhauled a little bit but may want to take a look at it later on. Which includes removing a variable that made and used for no reason and removing some other variables which weren't entirely needed.